moved DocumentTestsBase -> common
